2nd part of FS#12176. Tempo setting migrated to fixed point for libgme.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30274 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libgme/ay_emu.h
@@ -65,7 +65,7 @@ struct Ay_Emu {
 	int max_initial_silence;
 	int voice_count;
 	int mute_mask_;
-	double tempo;
+	int tempo;
 	double gain;
 	
 	long sample_rate;
@@ -141,7 +141,7 @@ long Track_get_length( struct Ay_Emu* this, int n );
 
 // Adjust song tempo, where 1.0 = normal, 0.5 = half speed, 2.0 = double speed.
 // Track length as returned by track_info() assumes a tempo of 1.0.
-void Sound_set_tempo( struct Ay_Emu* this, double t );
+void Sound_set_tempo( struct Ay_Emu* this, int t );
 	
 // Mute/unmute voice i, where voice 0 is first voice
 void Sound_mute_voice( struct Ay_Emu* this, int index, bool mute );
