+++ apps/codecs/libgme/gbs_emu.h
@@ -64,7 +64,7 @@ struct Gbs_Emu {
 	unsigned buf_changed_count;
 	int voice_count_;
 	double gain_;
-	double tempo_;
+	int tempo_;
 	
 	// track-specific
 	byte track_count;
@@ -158,7 +158,7 @@ static inline long Track_get_length( struct Gbs_Emu* this, int n )
 // Sound customization
 // Adjust song tempo, where 1.0 = normal, 0.5 = half speed, 2.0 = double speed.
 // Track length as returned by track_info() assumes a tempo of 1.0.
-void Sound_set_tempo( struct Gbs_Emu* this, double );
+void Sound_set_tempo( struct Gbs_Emu* this, int );
 
 // Mute/unmute voice i, where voice 0 is first voice
 void Sound_mute_voice( struct Gbs_Emu* this, int index, bool mute );
